not sending empty lines

+++ wrapper4nuclei_new_subs.sh
@@ -6,8 +6,12 @@ filteredSubsPath="/tmp/filtered_new_subs_nuclei.$(date +%s)"
 
 printf $subdomains2send | grep -vf /home/ec2-user/.config/cronScripts/nucleiBlacklist.txt > $filteredSubsPath
 
-message={\"domains\":\"$(cat $filteredSubsPath)\"}
+if [ -s "$filteredSubsPath" ]; then
+    echo "not empty" 
+    message={\"domains\":\"$(cat $filteredSubsPath)\"}
+    echo $message | sed 's/ /\\n/g' > $filteredSubsPath.message
+    GO111MODULE=off go run /home/ec2-user/rabbitmq/send.go -messageFile $filteredSubsPath.message -priority 99 -queue nuclei
+else
+  echo "$filteredSubsPath is empty after filtering blacklist"
+fi
 
-echo $message | sed 's/ /\\n/g' > $filteredSubsPath.message
-
-GO111MODULE=off go run /home/ec2-user/rabbitmq/send.go -messageFile $filteredSubsPath.message -priority 99 -queue nuclei
